\docType{data}
\name{cft_nutrients}
\alias{cft_nutrients}
\title{Nutritional composition and dry weight of crop functional types}
\description{
  Nutritional composition and dry weight of crop functional
  types
}
\references{
  see FAO (2001) and
  \url{http://www.fao.org/docrep/003/x9892e/X9892e05.htm#P8217_125315}
  for kcal, proteins and fat; Wirsenius (2000) for dry
  weights.
}
\keyword{crop}
\keyword{data,}
\keyword{datasets}
\keyword{dry}
\keyword{functional}
\keyword{nutrients,}
\keyword{types}
\keyword{weight,}

