\docType{data}
\name{CLUlonglat}
\alias{CLUlonglat}
\title{Exemplary CLUMondo Land Use Systems Map in longlat projection}
\description{
The SpatialPointsDataFrame includeds CLUMondo Land Use
Systems of the year 2040 for original CLUMondo grid cells
but in longlat projection.
}
\references{
A full documentation of the data can be found in: Van
Asselen, S., Verburg, P.H. (2012) A Land System
representation for global assessments and land-use
modeling. Global Change Biology. DOI:
10.1111/j.1365-2486.2012.02759.x.
}
\keyword{CLUMondo}
\keyword{cells,}
\keyword{grid}
\keyword{grid,}

