\docType{data}
\name{CLUMosaics}
\alias{CLUMosaics}
\title{Mosaic Composition of CLUMondo Land Use Systems}
\description{
  The array contains the composition of the 30 CLUMondo
  land use systems for all 24 CLUMondo world regions. The
  dimension of the array is [30, 8, 24], whereby there are
  30 Land use systems according to Asselen & Verburg (2012,
  2013); and the mosaics are comprised of 8 properties
  which refer to an pixel area of 9.25km x 9.25km =
  85.56km^2. Therefore, the areas are given as relative
  numbers. \itemize{ \item 0 = crop production (tons),
  \item 1 = livestock, bovines, goats and sheep (nr) \item
  2 = livestock, pigs and poultry (nr) \item 3 = builtup
  area (-) \item 4 = crop area (-) \item 5 = pasture area
  (-) \item 6 = tree area (-) \item 7 = bare area (-) }
}
\references{
  Van Asselen, S., Verburg, P.H. (2012) A Land System
  representation for global assessments and land-use
  modeling. Global Change Biology. DOI:
  10.1111/j.1365-2486.2012.02759.x.

  Van Asselen, S. & Verburg, P. H. (2013). Land cover
  change or land use intensification: simulting land system
  change with a global-scale land change model. Global
  Change Biology, 19, 3648-3667. DOI: 10.1111/gcb.12331
}
\keyword{CLUMondo,}
\keyword{datasets}
\keyword{land}
\keyword{mosaics,}
\keyword{systems}
\keyword{use}

