\docType{data}
\name{CLUtoLPJ2040}
\alias{CLUtoLPJ2040}
\title{Assignment of CLUMondo Land Use Systems to LPJmL Grid Cells for the year 2040}
\description{
  The data includes assignments of CLUMondoy Land Use
  Systems of the year 2040 to the old LPJmL grid,
  containing 59199 cells.
}
\keyword{cells,}
\keyword{CLUMondo}
\keyword{datasets}
\keyword{grid}
\keyword{grid,}
\keyword{LPJmL,}

