\docType{data}
\name{CLUtoLPJ2040long}
\alias{CLUtoLPJ2040long}
\title{Assignment of CLUMondo Land Use Systems to LPJmL Grid Cells for the 2040 OEPC scenario}
\description{
The data includes assignments of CLUMondoy Land Use Systems
of the 2040 OEPC scenario to new LPJmL grid, containing
67420 cells.
}
\references{
A full documentation of the data can be found in: Van
Asselen, S., Verburg, P.H. (2012) A Land System
representation for global assessments and land-use
modeling. Global Change Biology. DOI:
10.1111/j.1365-2486.2012.02759.x.
}
\keyword{CLUMondo}
\keyword{LPJmL,}
\keyword{cells,}
\keyword{grid}
\keyword{grid,}

