\docType{data}
\name{lpj_ingrid}
\alias{lpj_ingrid}
\title{Coordinates of LPJmL input grid cells identical to new output grid cells.}
\description{
The data provides coordinates of LPJmL input and output
grid cells. It includes 67420 cells with land use
information.
}
\keyword{LPJmL,}
\keyword{cells,}
\keyword{grid}

