\docType{data}
\name{lpj_long_clupos}
\alias{lpj_long_clupos}
\title{Positions of LPJmL output grid cells in the input grid}
\description{
  The data provides positions of LPJmL output grid cells
  within the input grid used for CLUMondo coupling. It
  includes 67420 land cells of the new LPJmL version.
}
\keyword{cells,}
\keyword{CLUMondo}
\keyword{datasets}
\keyword{grid}
\keyword{grid,}
\keyword{LPJmL,}

