\docType{data}
\name{lpj_short_clupos}
\alias{lpj_short_clupos}
\title{Positions of LPJmL output grid cells in the input grid}
\description{
The data provides positions of LPJmL output grid cells
within the input grid used for CLUMondo coupling. It
includes 59199 land cells of an older LPJmL version which
had a reduced number of output cells due to no soil
information for some cells.
}
\keyword{CLUMondo}
\keyword{LPJmL,}
\keyword{cells,}
\keyword{grid}
\keyword{grid,}

