\docType{data}
\name{lpj_short_outgrid}
\alias{lpj_short_outgrid}
\title{Coordinates of LPJmL output grid cells}
\description{
  The data provides coordinates of LPJmL output grid cells.
  It includes 59199 land cells of an older LPJmL version
  which had a reduced number of output cells due to no soil
  information for some cells.
}
\keyword{LPJmL,}
\keyword{cells,}
\keyword{datasets}
\keyword{grid}

