\docType{data}
\name{lpjGrid}
\alias{lpjGrid}
\title{Coordinates of LPJmL grid cells and information about countries and world regions.}
\description{
The SpatialPointsDataFrame provides coordinates of LPJmL
grid cells. It includes 67420 cells with information on the
country, region and corresponding CLUMondo world region.
}
\keyword{LPJmL,}
\keyword{cells,}
\keyword{grid}

