\docType{package}
\name{luess}
\alias{luess}
\alias{luess-package}
\title{Estimation of Ecosystem Services from LPJmL Model Output}
\description{
  Collection of methods to derive quantification of
  ecosystem services from LPJmL model output. The package
  also provides functions to translate CLUMondo land use
  system maps to LPJmL landuse input files. There are also
  some functions for easier LPJmL data handling and
  plotting which might be redundant to other
  \pkg{lu}-packages.
}
\author{
  Rene Sachse \email{rene.sachse@uni-potsdam.de}
}

