% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/resample_value.R
\name{resample_value}
\alias{resample_value}
\title{Resampling by calculating the means of covered cell centers}
\usage{
resample_value(grid_hr, grid_lr = generate_grid(), cells = lpj_long_clupos)
}
\arguments{
\item{grid_hr}{spatialPointsDataFrame of high resultion with the value of interest in first column}

\item{grid_lr}{regular grid, used for resampling}

\item{cells}{vector of integers; indicating which cells of the the grid_lr should be returned. By that also the order of the output cells can be determined.}
}
\value{
returns a SpatialPointsDataFrame
}
\description{
The function reads a spatialPointsDataFrame of high resolution
and a regular coarser grid. It than resamples by calculating the
means of all covered cell centers within the grid cell of the
coarser grid. The function will read only the first column of the
data.frame.
}
\author{
Rene Sachse \email{rene.sachse@uni-potsdam.de}
}
\seealso{
\code{\link[sp]{read.asciigrid}, \link[sp]{spTransform}} in package \pkg{sp}
}
\keyword{resampling}

