\docType{data}
\name{cft_nutrients}
\alias{cft_nutrients}
\title{Nutritional composition and dry weight of crop functional types}
\description{
The data set provides kcal per g WW, g protein per g WW, g
fat per g WW and DW/WW [\%] for the 16 crop functional
types.
}
\references{
See FAO (2001) and
\url{http://www.fao.org/docrep/003/x9892e/X9892e05.htm#P8217_125315}
for kcal, proteins and fat; Wirsenius (2000) for dry
weights.

Food and Agriculture Organization of the United Nations,
Rome, 2001. FOOD BALANCE SHEETS. A handbook.

Wirsenius, S., 2000. Human use of Land and Organic
Materials: Modeling the Turnover of Biomass in the Global
Food System. Dept. of Physical Resource Theory, Chalmers
University of Technology and Goeteborg University,
Goeteborg.
}
\keyword{crop}
\keyword{data,}
\keyword{dry}
\keyword{functional}
\keyword{nutrients,}
\keyword{types}
\keyword{weight,}

