\docType{package}
\name{luess}
\alias{luess}
\alias{luess-package}
\title{Estimation of Ecosystem Services from LPJmL Model Output}
\description{
  Collection of methods to derive quantification of
  ecosystem services from LPJmL model output.
}
\author{
  Rene Sachse <rene.sachse@uni-potsdam.de>
}

