\docType{data}
\name{smallarea}
\alias{smallarea}
\title{Artificial Landu Use Systems Map for testing purposes}
\description{
  The SpatialPointsDataFrame includeds cells with 6
  different artificial land use systems and is used for
  testing and validation purposes.
}
\keyword{cells,}
\keyword{CLUMondo}
\keyword{datasets}
\keyword{grid}
\keyword{grid,}

