\docType{data}
\name{smallarea}
\alias{smallarea}
\title{Artificial Land Use Systems Map for Testing Purposes}
\description{
The SpatialPointsDataFrame includes cells with 6 different
artificial land use systems and is used for testing and
validation purposes.
}
\keyword{CLUMondo}
\keyword{cells,}
\keyword{grid}
\keyword{grid,}

